+++ package.json
@@ -32,7 +32,7 @@
     "react": "^0.14.3",
     "react-dom": "^0.14.3",
     "react-redux": "4.3.0",
-    "react-router": "^2.0.1",
+    "react-router": "^4.0.0",
     "redux": "^3.0.4"
   }
 }
